requirements.txt
torch
transformers
streamlit
lime
newsapi-python
pandas
scikit-learn
tqdm